extract iterset
