+++ src/rruleset.ts
@@ -2,8 +2,7 @@ import RRule from './rrule'
 import dateutil from './dateutil'
 import { includes } from './helpers'
 import IterResult from './iterresult'
-import { DateWithZone } from './datewithzone'
-import { iter } from './iter'
+import { iterSet } from './iterset'
 
 export default class RRuleSet extends RRule {
   public readonly _rrule: RRule[]
@@ -46,7 +45,14 @@ export default class RRuleSet extends RRule {
   }
 
   _iter (iterResult: IterResult) {
-    return iterSet(iterResult, this._rrule, this._exrule, this._rdate, this._exdate, this.tzid())
+    return iterSet(
+      iterResult,
+      this._rrule,
+      this._exrule,
+      this._rdate,
+      this._exdate,
+      this.tzid()
+    )
   }
 
   /**
@@ -64,6 +70,20 @@ export default class RRuleSet extends RRule {
   }
 
   /**
+   * Adds an EXRULE to the set
+   *
+   * @param {RRule}
+   */
+  exrule (rrule: RRule) {
+    if (!(rrule instanceof RRule)) {
+      throw new TypeError(String(rrule) + ' is not RRule instance')
+    }
+    if (!includes(this._exrule.map(String), String(rrule))) {
+      this._exrule.push(rrule)
+    }
+  }
+
+  /**
    * Adds an RDate to the set
    *
    * @param {Date}
@@ -75,20 +95,6 @@ export default class RRuleSet extends RRule {
     if (!includes(this._rdate.map(Number), Number(date))) {
       this._rdate.push(date)
       dateutil.sort(this._rdate)
-    }
-  }
-
-  /**
-   * Adds an EXRULE to the set
-   *
-   * @param {RRule}
-   */
-  exrule (rrule: RRule) {
-    if (!(rrule instanceof RRule)) {
-      throw new TypeError(String(rrule) + ' is not RRule instance')
-    }
-    if (!includes(this._exrule.map(String), String(rrule))) {
-      this._exrule.push(rrule)
     }
   }
 
@@ -163,84 +169,12 @@ export default class RRuleSet extends RRule {
    */
   clone (): RRuleSet {
     const rrs = new RRuleSet(!!this._cache)
-    let i
-    for (i = 0; i < this._rrule.length; i++) {
-      rrs.rrule(this._rrule[i].clone())
-    }
-    for (i = 0; i < this._rdate.length; i++) {
-      rrs.rdate(new Date(this._rdate[i].getTime()))
-    }
-    for (i = 0; i < this._exrule.length; i++) {
-      rrs.exrule(this._exrule[i].clone())
-    }
-    for (i = 0; i < this._exdate.length; i++) {
-      rrs.exdate(new Date(this._exdate[i].getTime()))
-    }
+
+    this._rrule.forEach(rule => rrs.rrule(rule.clone()))
+    this._exrule.forEach(rule => rrs.exrule(rule.clone()))
+    this._rdate.forEach(date => rrs.rdate(new Date(date.getTime())))
+    this._exdate.forEach(date => rrs.exdate(new Date(date.getTime())))
+
     return rrs
-  }
-}
-
-function iterSet (iterResult: IterResult, _rrule: RRule[], _exrule: RRule[], _rdate: Date[], _exdate: Date[], tzid: string | undefined) {
-  const _exdateHash: { [k: number]: boolean } = {}
-  const _accept = iterResult.accept
-
-  function evalExdate (after: Date, before: Date) {
-    _exrule.forEach(function (rrule) {
-      rrule.between(after, before, true).forEach(function (date) {
-        _exdateHash[Number(date)] = true
-      })
-    })
-  }
-
-  _exdate.forEach(function (date) {
-    const zonedDate = new DateWithZone(date, tzid).rezonedDate()
-    _exdateHash[Number(zonedDate)] = true
-  })
-
-  iterResult.accept = function (date) {
-    const dt = Number(date)
-    if (!_exdateHash[dt]) {
-      evalExdate(new Date(dt - 1), new Date(dt + 1))
-      if (!_exdateHash[dt]) {
-        _exdateHash[dt] = true
-        return _accept.call(this, date)
-      }
-    }
-    return true
-  }
-
-  if (iterResult.method === 'between') {
-    evalExdate(iterResult.args.after!, iterResult.args.before!)
-    iterResult.accept = function (date) {
-      const dt = Number(date)
-      if (!_exdateHash[dt]) {
-        _exdateHash[dt] = true
-        return _accept.call(this, date)
-      }
-      return true
-    }
-  }
-
-  for (let i = 0; i < _rdate.length; i++) {
-    const zonedDate = new DateWithZone(_rdate[i], tzid).rezonedDate()
-    if (!iterResult.accept(new Date(zonedDate.getTime()))) break
-  }
-
-  _rrule.forEach(function (rrule) {
-    iter(iterResult, rrule.options)
-  })
-
-  const res = iterResult._result
-  dateutil.sort(res)
-  switch (iterResult.method) {
-    case 'all':
-    case 'between':
-      return res
-    case 'before':
-      return (res.length && res[res.length - 1]) || null
-    case 'after':
-      return (res.length && res[0]) || null
-    default:
-      return null
   }
 }
